JavaFX really sucks!
I cannot get it to work with JavaFX...
It might be smart that it's easy to make it look pretty, but it's really not smart how much code you have to write, to get the logic behind the "beautifil" GUI.

JavaFX sucks. And I will stop working with it from now on out. I can make a simple GUI for the Exams, and I can get some simple logic behind. But I refuse to work with this.